rename compound_sets to tasks
